refactor(app): :recycle: Added project support for golangci-lint: code refactored to placate the linter gods

+++ dms3server/server_config.go
@@ -20,7 +20,7 @@ type structSettings struct {
 // server details
 type structServer struct {
 	Port            int
-	CheckInterval   int
+	CheckInterval   uint16
 	EnableDashboard bool
 }
